Fix include directive for hdf5
In the original code HDF5_C_LIBRARY will only be looked for in the
local namespace, and not the ENV namespace since it is defined
as a variable. This patch reads from the ENV variable (which will
always exist since we explictly set it).

This is fixed in the upstream branch of netcdf-c.